+++ cmd/stack_output.go
@@ -17,8 +17,6 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/pulumi/pulumi/pkg/resource"
-
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -101,43 +99,6 @@ func newStackOutputCmd() *cobra.Command {
 	return cmd
 }
 
-// massagePropertyValue takes a property value and strips out the secrets annotations from it.  If showSecrets is
-// not true any secret values are replaced with "[secret]".
-func massagePropertyValue(v resource.PropertyValue, showSecrets bool) resource.PropertyValue {
-	switch {
-	case v.IsArray():
-		new := make([]resource.PropertyValue, len(v.ArrayValue()))
-		for i, e := range v.ArrayValue() {
-			new[i] = massagePropertyValue(e, showSecrets)
-		}
-		return resource.NewArrayProperty(new)
-	case v.IsObject():
-		new := make(resource.PropertyMap, len(v.ObjectValue()))
-		for k, e := range v.ObjectValue() {
-			new[k] = massagePropertyValue(e, showSecrets)
-		}
-		return resource.NewObjectProperty(massageSecrets(v.ObjectValue(), showSecrets))
-	case v.IsSecret() && showSecrets:
-		return massagePropertyValue(v.SecretValue().Element, showSecrets)
-	case v.IsSecret():
-		return resource.NewStringProperty("[secret]")
-	default:
-		return v
-	}
-}
-
-// massageSecrets takes a property map and returns a new map by transforming each value with massagePropertyValue
-// This allows us to serialize the resulting map using our existing serialization logic we use for deployments, to
-// produce sane output for stackOutputs.  If we did not do this, SecretValues would be serialized as objects
-// with the signature key and value.
-func massageSecrets(m resource.PropertyMap, showSecrets bool) resource.PropertyMap {
-	new := make(resource.PropertyMap, len(m))
-	for k, e := range m {
-		new[k] = massagePropertyValue(e, showSecrets)
-	}
-	return new
-}
-
 func getStackOutputs(snap *deploy.Snapshot, showSecrets bool) (map[string]interface{}, error) {
 	state, err := stack.GetRootStackResource(snap)
 	if err != nil {
@@ -147,5 +108,5 @@ func getStackOutputs(snap *deploy.Snapshot, showSecrets bool) (map[string]interf
 	// massageSecrets will remove all the secrets from the property map, so it should be safe to pass a panic
 	// crypter. This also ensure that if for some reason we didn't remove everything, we don't accidentally disclose
 	// secret values!
-	return stack.SerializeProperties(massageSecrets(state.Outputs, showSecrets), config.NewPanicCrypter())
+	return stack.SerializeProperties(display.MassageSecrets(state.Outputs, showSecrets), config.NewPanicCrypter())
 }
